CR refactor: rename implementations to Impl
