+++ std-bits/aws/src/main/java/org/enso/aws/file_system/S3DataLinkImpl.java
@@ -1,21 +1,21 @@
-package org.enso.database.postgres;
+package org.enso.aws.file_system;
 
 import org.enso.base.enso_cloud.DataLinkSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DataLinkSPI.class)
-public class PostgresDataLinkSPI extends DataLinkSPI {
+public class S3DataLinkImpl extends DataLinkSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.Database.Connection.Data_Link.Postgres_Data_Link";
+    return "Standard.AWS.S3.S3_Data_Link";
   }
 
   @Override
   protected String getTypeName() {
-    return "Postgres_Data_Link";
+    return "S3_Data_Link";
   }
 
   @Override
   protected String getLinkTypeName() {
-    return "Postgres_Connection";
+    return "S3";
   }
 }
