+++ std-bits/aws/src/main/java/org/enso/aws/file_system/S3FileSystemImpl.java
@@ -3,7 +3,7 @@ package org.enso.aws.file_system;
 import org.enso.base.file_system.FileSystemSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = FileSystemSPI.class)
-public class S3FileSystemSPI extends FileSystemSPI {
+public class S3FileSystemImpl extends FileSystemSPI {
   @Override
   protected String getModuleName() {
     return "Standard.AWS.S3.S3_File";
