+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/EnsoPathFileSystemImpl.java
@@ -8,7 +8,7 @@ import org.enso.base.file_system.FileSystemSPI;
  * <p>See `Enso_File.new` for more information on path resolution.
  */
 @org.openide.util.lookup.ServiceProvider(service = FileSystemSPI.class)
-public class EnsoPathFileSystemSPI extends FileSystemSPI {
+public class EnsoPathFileSystemImpl extends FileSystemSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.Enso_Cloud.Enso_File";
