+++ std-bits/base/src/main/java/org/enso/base/file_format/JSONFileFormatImpl.java
@@ -1,7 +1,7 @@
 package org.enso.base.file_format;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class ByteFormatSPI extends FileFormatSPI {
+public class JSONFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.System.File_Format";
@@ -9,6 +9,11 @@ public class ByteFormatSPI extends FileFormatSPI {
 
   @Override
   protected String getTypeName() {
-    return "Bytes";
+    return "JSON_Format";
+  }
+
+  @Override
+  protected String getDataLinkFormatName() {
+    return "json";
   }
 }
