+++ std-bits/base/src/main/java/org/enso/base/file_format/TextFileFormatImpl.java
@@ -1,7 +1,7 @@
 package org.enso.base.file_format;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class JSONFormatSPI extends FileFormatSPI {
+public class TextFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.System.File_Format";
@@ -9,11 +9,6 @@ public class JSONFormatSPI extends FileFormatSPI {
 
   @Override
   protected String getTypeName() {
-    return "JSON_Format";
-  }
-
-  @Override
-  protected String getDataLinkFormatName() {
-    return "json";
+    return "Plain_Text_Format";
   }
 }
