+++ std-bits/base/src/main/java/org/enso/base/file_format/XMLFileFormatImpl.java
@@ -1,7 +1,7 @@
 package org.enso.base.file_format;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class XMLFormatSPI extends FileFormatSPI {
+public class XMLFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.Data.XML.XML_Format";
