+++ std-bits/base/src/main/java/org/enso/base/net/http/HTTPFetchDataLinkImpl.java
@@ -3,7 +3,7 @@ package org.enso.base.net.http;
 import org.enso.base.enso_cloud.DataLinkSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DataLinkSPI.class)
-public class HTTPFetchDataLinkSPI extends DataLinkSPI {
+public class HTTPFetchDataLinkImpl extends DataLinkSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.Network.HTTP.Internal.HTTP_Fetch_Data_Link";
