+++ std-bits/base/src/main/java/org/enso/base/read/BaseReadManyReturnImpl.java
@@ -1,7 +1,7 @@
 package org.enso.base.read;
 
 @org.openide.util.lookup.ServiceProvider(service = ReadManyReturnSPI.class)
-public class BaseReadManyReturnSPI extends ReadManyReturnSPI {
+public class BaseReadManyReturnImpl extends ReadManyReturnSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Base.Data.Read.Return_As";
