+++ std-bits/database/src/main/java/org/enso/database/postgres/PostgresConnectionDetailsImpl.java
@@ -1,27 +1,26 @@
-package org.enso.microsoft;
+package org.enso.database.postgres;
 
 import org.enso.database.DatabaseConnectionDetailsSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DatabaseConnectionDetailsSPI.class)
-public class SQLServerConnectionDetailsSPI extends DatabaseConnectionDetailsSPI {
-
+public class PostgresConnectionDetailsImpl extends DatabaseConnectionDetailsSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.Microsoft.Connection.SQLServer_Details";
+    return "Standard.Database.Connection.Postgres";
   }
 
   @Override
   protected String getTypeName() {
-    return "SQLServer_Details";
+    return "Postgres";
   }
 
   @Override
   protected String getCodeForDefaultConstructor() {
-    return "..SQLServer";
+    return "(Postgres.Server 'localhost' 5432)";
   }
 
   @Override
   protected String getUserFacingConnectionName() {
-    return "Microsoft SQL Server";
+    return "Postgres";
   }
 }
