+++ std-bits/database/src/main/java/org/enso/database/postgres/PostgresDataLinkImpl.java
@@ -1,21 +1,21 @@
-package org.enso.aws.file_system;
+package org.enso.database.postgres;
 
 import org.enso.base.enso_cloud.DataLinkSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DataLinkSPI.class)
-public class S3DataLinkSPI extends DataLinkSPI {
+public class PostgresDataLinkImpl extends DataLinkSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.AWS.S3.S3_Data_Link";
+    return "Standard.Database.Connection.Data_Link.Postgres_Data_Link";
   }
 
   @Override
   protected String getTypeName() {
-    return "S3_Data_Link";
+    return "Postgres_Data_Link";
   }
 
   @Override
   protected String getLinkTypeName() {
-    return "S3";
+    return "Postgres_Connection";
   }
 }
