+++ std-bits/database/src/main/java/org/enso/database/sqlite/SQLiteConnectionDetailsImpl.java
@@ -1,26 +1,26 @@
-package org.enso.aws.database;
+package org.enso.database.sqlite;
 
 import org.enso.database.DatabaseConnectionDetailsSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DatabaseConnectionDetailsSPI.class)
-public class RedshiftConnectionDetailsSPI extends DatabaseConnectionDetailsSPI {
+public class SQLiteConnectionDetailsImpl extends DatabaseConnectionDetailsSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.AWS.Database.Redshift.Redshift_Details";
+    return "Standard.Database.Connection.SQLite";
   }
 
   @Override
   protected String getTypeName() {
-    return "Redshift_Details";
+    return "SQLite";
   }
 
   @Override
   protected String getCodeForDefaultConstructor() {
-    return "..Redshift";
+    return "SQLite.From_File";
   }
 
   @Override
   protected String getUserFacingConnectionName() {
-    return "Redshift";
+    return "SQLite";
   }
 }
