+++ std-bits/database/src/main/java/org/enso/database/sqlite/SQLiteFileFormatImpl.java
@@ -3,7 +3,7 @@ package org.enso.database.sqlite;
 import org.enso.base.file_format.FileFormatSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class SQLiteFormatSPI extends FileFormatSPI {
+public class SQLiteFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Database.Connection.SQLite_Format";
