+++ std-bits/database/src/main/java/org/enso/database/sqlite/SQLiteInMemoryDetailsImpl.java
@@ -3,7 +3,7 @@ package org.enso.database.sqlite;
 import org.enso.database.DatabaseConnectionDetailsSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DatabaseConnectionDetailsSPI.class)
-public class SQLiteConnectionDetailsSPI extends DatabaseConnectionDetailsSPI {
+public class SQLiteInMemoryDetailsImpl extends DatabaseConnectionDetailsSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Database.Connection.SQLite";
@@ -16,11 +16,11 @@ public class SQLiteConnectionDetailsSPI extends DatabaseConnectionDetailsSPI {
 
   @Override
   protected String getCodeForDefaultConstructor() {
-    return "SQLite.From_File";
+    return "SQLite.In_Memory";
   }
 
   @Override
   protected String getUserFacingConnectionName() {
-    return "SQLite";
+    return "SQLite (In-Memory)";
   }
 }
