+++ std-bits/snowflake/src/main/java/org/enso/snowflake/SnowflakeDataLinkImpl.java
@@ -3,7 +3,7 @@ package org.enso.snowflake;
 import org.enso.base.enso_cloud.DataLinkSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DataLinkSPI.class)
-public class SnowflakeDataLinkSPI extends DataLinkSPI {
+public class SnowflakeDataLinkImpl extends DataLinkSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Snowflake.Snowflake_Data_Link";
