+++ std-bits/table/src/main/java/org/enso/table/read/DelimitedFileFormatImpl.java
@@ -3,19 +3,19 @@ package org.enso.table.read;
 import org.enso.base.file_format.FileFormatSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class ExcelFormatSPI extends FileFormatSPI {
+public class DelimitedFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.Table.Excel.Excel_Format";
+    return "Standard.Table.Delimited.Delimited_Format";
   }
 
   @Override
   protected String getTypeName() {
-    return "Excel_Format";
+    return "Delimited_Format";
   }
 
   @Override
   protected String getDataLinkFormatName() {
-    return "excel";
+    return "delimited";
   }
 }
