+++ std-bits/table/src/main/java/org/enso/table/read/ExcelFileFormatImpl.java
@@ -3,19 +3,19 @@ package org.enso.table.read;
 import org.enso.base.file_format.FileFormatSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = FileFormatSPI.class)
-public class DelimitedFormatSPI extends FileFormatSPI {
+public class ExcelFileFormatImpl extends FileFormatSPI {
   @Override
   protected String getModuleName() {
-    return "Standard.Table.Delimited.Delimited_Format";
+    return "Standard.Table.Excel.Excel_Format";
   }
 
   @Override
   protected String getTypeName() {
-    return "Delimited_Format";
+    return "Excel_Format";
   }
 
   @Override
   protected String getDataLinkFormatName() {
-    return "delimited";
+    return "excel";
   }
 }
