+++ std-bits/table/src/main/java/org/enso/table/read/TableReadManyReturnImpl.java
@@ -3,7 +3,7 @@ package org.enso.table.read;
 import org.enso.base.read.ReadManyReturnSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = ReadManyReturnSPI.class)
-public class TableReadManyReturnSPI extends ReadManyReturnSPI {
+public class TableReadManyReturnImpl extends ReadManyReturnSPI {
   @Override
   protected String getModuleName() {
     return "Standard.Table.Return_As_Table";
